Cleanup file structure topology
